Each model has its own script for building datasets.
convert_sequences_to_melodies.py is turned into a library. basic_rnn has
create_dataset.py which imports sequence_to_melodies and builds the
dataset. Models will define their own encoders in their
create_dataset.py scripts. In this case, a default encoder is defined in
sequence_to_melodies.py and used by basic_rnn.
